Refactor CMake configuration and main logic; rename executable targets and add referee process
