+++ CMakeLists.txt
@@ -6,14 +6,14 @@ set(CMAKE_C_STANDARD 11)
 find_package(OpenGL REQUIRED)
 find_package(GLUT REQUIRED)
 
-add_executable(referee src/main.c src/utils/config.c
+add_executable(referee src/referee.c src/utils/config.c
 src/utils/random.c src/utils/player_utils.c src/utils/file.c src/utils/referee_orders.c)
 
 add_executable(player src/player.c src/utils/player_utils.c src/utils/random.c)
 
 add_executable(rope_game_graphics src/graphics/main_graphics.c)
 
-add_executable(rope_pulling_game_main src/main_temp.c)
+add_executable(rope_pulling_game_main src/main.c)
 
 target_link_libraries(rope_game_graphics
     PRIVATE
